+++ testsuites/psxtmtests/psxtmcond08/init.c
@@ -50,7 +50,7 @@ void *Low(
   void *argument
 )
 {
-  long end_time;
+  uint32_t end_time;
 
   end_time = benchmark_timer_read();
